@@ -25,24 +26,34 @@ apply plugin: 'idea'
 group = 'uk.gov.hmcts.reform'
 version = '0.0.1'
 
-allprojects {
-  sourceCompatibility = '11'
-  targetCompatibility = '11'
+java {
+  toolchain {
+    languageVersion = JavaLanguageVersion.of(17)
+  }
+}
+
+ext {
+  springCloudVersion = '2023.0.3'
 }
 
 def versions = [
-  lombok          : '1.18.20',
-  gradlePitest    : '1.5.1',
-  pitest          : '1.7.0',
-  sonarPitest     : '0.5',
-  reformLogging   : '5.1.7',
-  springfoxSwagger: '2.9.2',
-  pact_version    : '4.1.7',
-  log4j           : '2.17.1',
-  launchDarklySdk : "5.2.3",
-  reformS2sClient : '4.0.0',
-  serenity        : '2.0.76',
-  restAssured     : '4.3.3'
+        lombok          : '1.18.32',
+        gradlePitest    : '1.5.1',
+        pitest          : '1.16.1',
+        sonarPitest     : '0.5',
+        reformLogging   : '6.1.4',
+        pact_version    : '4.1.7',
+        log4j           : '2.23.1',
+        launchDarklySdk : "5.10.9",
+        reformS2sClient : '5.1.0',
+        springHystrix   : '2.0.76',
+        restAssured     : '4.3.3',
+        feign           : '3.8.0',
+        bytebuddy       : '1.14.18',
+        testContainer_postgresql: '1.19.8',
+        junit           : '5.9.2',
+        junitPlatform   : '1.10.3',
+        bouncycastle    : '1.78.1'
 ]
 
 sourceSets {
