@@ -1,14 +1,19 @@
 package uk.gov.hmcts.reform.cdapi.config;
 
+import jakarta.inject.Inject;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.annotation.Order;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.oauth2.client.web.OAuth2AuthorizationRequestRedirectFilter;
 import org.springframework.security.oauth2.core.DelegatingOAuth2TokenValidator;
 import org.springframework.security.oauth2.core.OAuth2TokenValidator;
 import org.springframework.security.oauth2.jwt.Jwt;
@@ -17,22 +22,19 @@ import org.springframework.security.oauth2.jwt.JwtDecoders;
 import org.springframework.security.oauth2.jwt.JwtTimestampValidator;
 import org.springframework.security.oauth2.jwt.NimbusJwtDecoder;
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationConverter;
-import org.springframework.security.oauth2.server.resource.web.BearerTokenAuthenticationFilter;
+import org.springframework.security.oauth2.server.resource.web.authentication.BearerTokenAuthenticationFilter;
+import org.springframework.security.web.SecurityFilterChain;
 import uk.gov.hmcts.reform.authorisation.filters.ServiceAuthFilter;
 import uk.gov.hmcts.reform.cdapi.oidc.JwtGrantedAuthoritiesConverter;
 
 import java.util.List;
-import javax.inject.Inject;
-
-import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
 
 @Configuration
 @ConfigurationProperties(prefix = "security")
 @EnableWebSecurity
 @Slf4j
 @SuppressWarnings("unchecked")
-public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
-
+public class SecurityConfiguration {
 
     @Value("${spring.security.oauth2.client.provider.oidc.issuer-uri}")
     private String issuerUri;
@@ -40,7 +42,10 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
     @Value("${oidc.issuer}")
     private String issuerOverride;
 
-    private ServiceAuthFilter serviceAuthFilter;
+    @Order(1)
+    private  ServiceAuthFilter serviceAuthFilter;
+    @Order(2)
+    private final SecurityEndpointFilter securityEndpointFilter;
     List<String> anonymousPaths;
 
     private JwtAuthenticationConverter jwtAuthenticationConverter;
@@ -55,43 +60,37 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
         this.anonymousPaths = anonymousPaths;
     }
 
-    @Override
-    public void configure(WebSecurity web) {
-        web.ignoring()
-            .antMatchers(anonymousPaths.toArray(new String[0]));
+    @Bean
+    public WebSecurityCustomizer webSecurityCustomizer() {
+        return web -> web.ignoring().requestMatchers(anonymousPaths.toArray(String[]::new));
     }
 
     @Inject
     public SecurityConfiguration(final JwtGrantedAuthoritiesConverter jwtGrantedAuthoritiesConverter,
                                  final ServiceAuthFilter serviceAuthFilter,
-                                 RestAuthenticationEntryPoint restAuthenticationEntryPoint) {
+                                 RestAuthenticationEntryPoint restAuthenticationEntryPoint,
+                                 SecurityEndpointFilter securityEndpointFilter) {
 
         this.serviceAuthFilter = serviceAuthFilter;
         this.restAuthenticationEntryPoint = restAuthenticationEntryPoint;
         jwtAuthenticationConverter = new JwtAuthenticationConverter();
         jwtAuthenticationConverter.setJwtGrantedAuthoritiesConverter(jwtGrantedAuthoritiesConverter);
-
+        this.securityEndpointFilter = securityEndpointFilter;
     }
 
-    @Override
-    protected void configure(final HttpSecurity http) throws Exception {
-        http
-            .addFilterBefore(serviceAuthFilter, BearerTokenAuthenticationFilter.class)
-            .sessionManagement().sessionCreationPolicy(STATELESS).and()
-            .csrf().disable()
-            .formLogin().disable()
-            .logout().disable()
-            .authorizeRequests()
-            .antMatchers("/error").permitAll()
-            .anyRequest()
-            .authenticated()
-            .and()
-            .oauth2ResourceServer().authenticationEntryPoint(restAuthenticationEntryPoint)
-            .jwt()
-            .jwtAuthenticationConverter(jwtAuthenticationConverter)
-            .and()
-            .and()
-            .oauth2Client();
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        http.addFilterBefore(serviceAuthFilter, BearerTokenAuthenticationFilter.class)
+            .addFilterAfter(securityEndpointFilter, OAuth2AuthorizationRequestRedirectFilter.class)
+            .sessionManagement(s -> s.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+            .csrf(AbstractHttpConfigurer::disable)
+            .formLogin(AbstractHttpConfigurer::disable)
+            .logout(AbstractHttpConfigurer::disable)
+            .authorizeHttpRequests(a -> a.requestMatchers("/error").permitAll().anyRequest().authenticated())
+            .oauth2ResourceServer(a -> a.authenticationEntryPoint(restAuthenticationEntryPoint)
+                .jwt(j -> j.jwtAuthenticationConverter(jwtAuthenticationConverter)))
+            .oauth2Client(Customizer.withDefaults());
+        return http.build();
     }
 
     @Bean
