@@ -11,9 +11,10 @@ import org.springframework.test.context.ActiveProfiles;
 import uk.gov.hmcts.reform.cdapi.controllers.response.Categories;
 import uk.gov.hmcts.reform.cdapi.domain.CaseFlag;
 import uk.gov.hmcts.reform.cdapi.exception.ErrorResponse;
-import uk.gov.hmcts.reform.cdapi.serenity5.SerenityTest;
+import uk.gov.hmcts.reform.cdapi.util.ErrorInvalidRequestResponse;
 import uk.gov.hmcts.reform.cdapi.util.FeatureToggleConditionExtension;
 import uk.gov.hmcts.reform.cdapi.util.ToggleEnable;
+import uk.gov.hmcts.reform.lib.util.serenity5.SerenityTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -35,8 +36,14 @@ class CommonDataApiFunctionalTest extends AuthorizationFunctionalTest {
     private static final String MAP_KEY_LOV = "CrdApiController.retrieveListOfValuesByCategoryId";
     private static final String PATH_LOV = SLASH.concat("lov").concat(SLASH).concat("categories");
     private static final String PARAM_SIGN = "?";
-    private static final String PARAM_HEARING = SLASH.concat("HearingChannel");
+    private static final String PARAM_HEARING = SLASH.concat("HearingSubChannel");
     private static final String PARAM_HEARING_WITH_PARAM_SIGN = PARAM_HEARING.concat(PARAM_SIGN);
+    private static final String SERVICE_ID = "ServiceId=";
+    private static final String SERVICE_ID_BBA3 = "BBA3";
+    private static final String PARAM_LISTING_STATUS_WITH_SERVICE_ID = SLASH.concat("ListingStatus")
+        .concat(PARAM_SIGN).concat(SERVICE_ID + SERVICE_ID_BBA3);
+    private static final String PARAM_LISTING_STATUS_WITH_EMPTY_SERVICE_ID = SLASH.concat("ListingStatus")
+        .concat(PARAM_SIGN).concat(SERVICE_ID);
     private static final String DATA_NOT_FOUND = "Data not found";
 
     @Test
@@ -46,7 +53,7 @@ class CommonDataApiFunctionalTest extends AuthorizationFunctionalTest {
         final var response = (ErrorResponse)
             commonDataApiClient.retrieveCaseFlagsByServiceId(
                 HttpStatus.BAD_REQUEST,
-                "hello"
+                "service-id=XXXX?flag-type=hello"
             );
         assertNotNull(response);
         assertEquals("Allowed values are PARTY or CASE", response.getErrorDescription());
@@ -109,8 +116,8 @@ class CommonDataApiFunctionalTest extends AuthorizationFunctionalTest {
         if (OK.value() == response.getStatusCode()) {
             var categories = response.getBody().as(Categories.class);
             assertNotNull(categories);
-            assertThat(categories.getListOfCategory()).hasSizeGreaterThan(1);
-            categories.getListOfCategory().forEach(h -> assertEquals("HearingChannel", h.getCategoryKey()));
+            assertThat(categories.getListOfCategory()).hasSizeGreaterThan(0);
+            categories.getListOfCategory().forEach(h -> assertEquals("HearingSubChannel", h.getCategoryKey()));
             categories.getListOfCategory().forEach(h -> assertNull(h.getChildNodes()));
         } else {
             assertEquals(NOT_FOUND.value(), response.getStatusCode());
@@ -130,7 +137,7 @@ class CommonDataApiFunctionalTest extends AuthorizationFunctionalTest {
             var categories = response.getBody().as(Categories.class);
             assertNotNull(categories);
             assertThat(categories.getListOfCategory()).hasSizeGreaterThan(0);
-            categories.getListOfCategory().forEach(h -> assertEquals("HearingChannel", h.getCategoryKey()));
+            categories.getListOfCategory().forEach(h -> assertEquals("HearingSubChannel", h.getCategoryKey()));
             categories.getListOfCategory().forEach(h -> assertFalse(h.getChildNodes().isEmpty()));
         } else {
             assertEquals(NOT_FOUND.value(), response.getStatusCode());
@@ -166,7 +173,7 @@ class CommonDataApiFunctionalTest extends AuthorizationFunctionalTest {
         if (OK.value() == response.getStatusCode()) {
             var categories = response.getBody().as(Categories.class);
             assertNotNull(categories);
-            categories.getListOfCategory().forEach(h -> assertEquals("HearingChannel", h.getCategoryKey()));
+            categories.getListOfCategory().forEach(h -> assertEquals("HearingSubChannel", h.getCategoryKey()));
             assertThat(categories.getListOfCategory()).hasSizeGreaterThan(0);
         } else {
             assertEquals(NOT_FOUND.value(), response.getStatusCode());
@@ -195,7 +202,7 @@ class CommonDataApiFunctionalTest extends AuthorizationFunctionalTest {
             commonDataApiClient.retrieveListOfValuesByCategoryId(
                 NOT_FOUND,
                 PARAM_HEARING_WITH_PARAM_SIGN
-                    + "isChildRequired=y&parentCategory=HearingChannel"
+                    + "isChildRequired=y&parentCategory=HearingSubChannel"
             );
         assertNotNull(response);
         assertEquals(DATA_NOT_FOUND, response.getErrorDescription());
@@ -253,12 +260,102 @@ class CommonDataApiFunctionalTest extends AuthorizationFunctionalTest {
     @ExtendWith(FeatureToggleConditionExtension.class)
     @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
     void shouldThrowErrorWhenCategoryIdIsEmpty() {
-        final var response = (ErrorResponse)
+        final var response = (ErrorInvalidRequestResponse)
             commonDataApiClient.retrieveBadRequestByEmptyCategoryId(
-                HttpStatus.BAD_REQUEST,
-                " "
+                NOT_FOUND,
+                ""
             );
         assertNotNull(response);
-        assertEquals("Syntax error or Bad request", response.getErrorDescription());
+        assertEquals("Not Found", response.getError());
     }
+
+    @Test
+    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
+    @ExtendWith(FeatureToggleConditionExtension.class)
+    void shouldReturnChildCategoriesInParticularToServiceId() {
+        Response response =
+            commonDataApiClient.retrieveCategoriesByCategoryIdSuccess(PATH_LOV, "/HearingSubChannel?"
+                + "isChildRequired=y&serviceId=BBA3&key=VID");
+
+        if (OK.value() == response.getStatusCode()) {
+            var categories = response.getBody().as(Categories.class);
+            assertNotNull(categories);
+            categories.getListOfCategory().forEach(h -> assertEquals("HearingSubChannel", h.getCategoryKey()));
+            assertThat(categories.getListOfCategory()).hasSizeGreaterThan(0);
+            assertThat(categories.getListOfCategory().get(0).getChildNodes()).hasSizeGreaterThan(0);
+            assertEquals("HearingSubChannel", categories.getListOfCategory().get(0).getChildNodes().get(0)
+                .getParentCategory());
+
+        } else {
+            assertEquals(NOT_FOUND.value(), response.getStatusCode());
+        }
+    }
+
+    @Test
+    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
+    @ExtendWith(FeatureToggleConditionExtension.class)
+    void shouldReturnSuccessWithServiceIdNoChilds() {
+        Response response =
+            commonDataApiClient.retrieveCategoriesByCategoryIdSuccess(PATH_LOV, PARAM_HEARING_WITH_PARAM_SIGN
+                + "isChildRequired=n&serviceId=BBA3");
+
+        if (OK.value() == response.getStatusCode()) {
+            var categories = response.getBody().as(Categories.class);
+            assertNotNull(categories);
+            categories.getListOfCategory().forEach(h -> assertEquals("HearingSubChannel", h.getCategoryKey()));
+            assertThat(categories.getListOfCategory()).hasSizeGreaterThan(0);
+            categories.getListOfCategory().forEach(h -> assertNull(h.getChildNodes()));
+        } else {
+            assertEquals(NOT_FOUND.value(), response.getStatusCode());
+        }
+    }
+
+    @Test
+    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
+    @ExtendWith(FeatureToggleConditionExtension.class)
+    void shouldReturnSuccess_Valid_ServiceID() {
+        Response response =
+            commonDataApiClient.retrieveCategoriesByCategoryIdSuccess(PATH_LOV, PARAM_LISTING_STATUS_WITH_SERVICE_ID);
+        if (OK.value() == response.getStatusCode()) {
+            var categories = response.getBody().as(Categories.class);
+            assertNotNull(categories);
+            assertThat(categories.getListOfCategory()).hasSizeGreaterThan(1);
+            categories.getListOfCategory().forEach(h -> assertEquals("ListingStatus", h.getCategoryKey()));
+        } else {
+            assertEquals(NOT_FOUND.value(), response.getStatusCode());
+        }
+    }
+
+    @Test
+    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
+    @ExtendWith(FeatureToggleConditionExtension.class)
+    void shouldReturnSuccess_Empty_ServiceID() {
+        Response response =
+            commonDataApiClient.retrieveCategoriesByCategoryIdSuccess(
+                PATH_LOV,
+                PARAM_LISTING_STATUS_WITH_EMPTY_SERVICE_ID
+            );
+        if (OK.value() == response.getStatusCode()) {
+            var categories = response.getBody().as(Categories.class);
+            assertNotNull(categories);
+            assertThat(categories.getListOfCategory()).hasSizeGreaterThan(1);
+            categories.getListOfCategory().forEach(h -> assertEquals("ListingStatus", h.getCategoryKey()));
+        } else {
+            assertEquals(NOT_FOUND.value(), response.getStatusCode());
+        }
+    }
+
+    @Test
+    @ToggleEnable(mapKey = MAP_KEY_CASE_FLAGS, withFeature = true)
+    @ExtendWith(FeatureToggleConditionExtension.class)
+    void shouldReturnBadRequestForRetrieveCaseFlagsByServiceIdWithAvailableExternalFlagIsEmpty() {
+        final var response = (ErrorResponse)
+            commonDataApiClient.retrieveCaseFlagsByServiceId(
+                HttpStatus.BAD_REQUEST,
+                "service-id=XXXX?available-external-flag="
+            );
+        assertEquals(response.getErrorCode(), 400);
+        assertEquals(response.getErrorDescription(), "Allowed values are Y or N");
+    }
+
 }
