@@ -4,11 +4,12 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
+import uk.gov.hmcts.reform.lib.config.TestConfig;
 
 @Getter
 @Setter
 @Configuration
-public class TestConfigProperties {
+public class TestConfigProperties implements TestConfig {
 
 
     @Value("${oauth2.client.secret}")
@@ -34,5 +35,8 @@ public class TestConfigProperties {
 
     @Value("${targetInstance}")
     protected String commonDataApiUrl;
+
+    @Value("${scope-name}")
+    protected String scope;
 
 }
