@@ -4,30 +4,21 @@ import com.google.gson.Gson;
 import com.google.gson.annotations.SerializedName;
 import com.mifmif.common.regex.Generex;
 import io.restassured.RestAssured;
-import io.restassured.response.Response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import uk.gov.hmcts.reform.cdapi.config.TestConfigProperties;
+import uk.gov.hmcts.reform.lib.idam.IdamOpenId;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.UUID;
 
 import static org.apache.commons.lang3.StringUtils.isBlank;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
 import static org.springframework.http.MediaType.APPLICATION_FORM_URLENCODED_VALUE;
-import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-import static uk.gov.hmcts.reform.cdapi.AuthorizationFunctionalTest.CREDS;
-import static uk.gov.hmcts.reform.cdapi.AuthorizationFunctionalTest.EMAIL;
-import static uk.gov.hmcts.reform.cdapi.AuthorizationFunctionalTest.generateRandomEmail;
+
 
 @Slf4j
-public class IdamOpenIdClient {
+public class IdamOpenIdClient extends IdamOpenId {
 
     private TestConfigProperties testConfig;
 
@@ -38,93 +29,9 @@ public class IdamOpenIdClient {
     private static String sidamPassword;
 
     public IdamOpenIdClient(TestConfigProperties testConfig) {
+        super(testConfig);
         this.testConfig = testConfig;
     }
-
-    public Map<String, String> createUser(String userRole) {
-        return createUser(userRole, generateRandomEmail(), "First", "Last");
-    }
-
-    public Map<String, String> createUser(String userRole, String userEmail, String firstName, String lastName) {
-        //Generating a random user
-        String userGroup = "";
-        String password = generateSidamPassword();
-
-        String id = UUID.randomUUID().toString();
-
-        Role role = new Role(userRole);
-
-        List<Role> roles = new ArrayList<>();
-        roles.add(role);
-
-        Group group = new Group(userGroup);
-
-        User user = new User(userEmail, firstName, id, lastName, password, roles, group);
-
-        String serializedUser = gson.toJson(user);
-
-        log.info("Creating new Sidam user");
-        Response createdUserResponse = RestAssured
-            .given()
-            .relaxedHTTPSValidation()
-            .baseUri(testConfig.getIdamApiUrl())
-            .header(CONTENT_TYPE, APPLICATION_JSON_VALUE)
-            .body(serializedUser)
-            .post("/testing-support/accounts")
-            .andReturn();
-
-
-        log.info("openIdTokenResponse createUser response: " + createdUserResponse.getStatusCode());
-
-        assertThat(createdUserResponse.getStatusCode()).isEqualTo(201);
-
-        Map<String, String> userCreds = new HashMap<>();
-        userCreds.put(EMAIL, userEmail);
-        userCreds.put(CREDS, password);
-        return userCreds;
-    }
-
-    public String getOpenIdToken() {
-
-        if (openIdTokenCommonDataAdmin == null) {
-            Map<String, String> userCreds = createUser("commondata-admin");
-            openIdTokenCommonDataAdmin = getOpenIdToken(userCreds.get(EMAIL), userCreds.get(CREDS));
-        }
-        return openIdTokenCommonDataAdmin;
-    }
-
-    public String getOpenIdToken(String userEmail, String password) {
-
-
-        Map<String, String> tokenParams = new HashMap<>();
-        tokenParams.put("grant_type", "password");
-        tokenParams.put("username", userEmail);
-        tokenParams.put("password", password);
-        tokenParams.put("client_id", testConfig.getClientId());
-        tokenParams.put("client_secret", testConfig.getClientSecret());
-        tokenParams.put("redirect_uri", testConfig.getOauthRedirectUrl());
-        tokenParams.put("scope", "openid profile roles search-user");
-
-        log.info("Creating Sidam token");
-        Response openIdTokenResponse = RestAssured
-            .given()
-            .relaxedHTTPSValidation()
-            .baseUri(testConfig.getIdamApiUrl())
-            .header(CONTENT_TYPE, APPLICATION_FORM_URLENCODED_VALUE)
-            .params(tokenParams)
-            .post("/o/token")
-            .andReturn();
-
-        log.info("getOpenIdToken response: " + openIdTokenResponse.getStatusCode());
-
-        assertEquals(200, openIdTokenResponse.getStatusCode());
-
-        BearerTokenResponse accessTokenResponse = gson.fromJson(openIdTokenResponse.getBody()
-                                                                    .asString(), BearerTokenResponse.class);
-        return accessTokenResponse.getAccessToken();
-
-    }
-
 
     public void deleteSidamUser(String email) {
         try {
