@@ -12,7 +12,7 @@ import org.springframework.context.annotation.PropertySource;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
-import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.HttpStatusCodeException;
 import org.springframework.web.client.RestTemplate;
@@ -58,9 +58,15 @@ public class CommonDataApiClient {
         return mapCaseFlagsByServiceIdResponse(responseEntity, clazz);
     }
 
+    public Object retrieveCaseFlagsByServiceIdJsonFormat(String queryParam, Class<?> clazz,
+                                                         String path) {
+        ResponseEntity<Object> responseEntity = getRequest(APP_BASE_PATH + path + queryParam, String.class, "");
+        return responseEntity.getBody();
+    }
+
     private Object mapCaseFlagsByServiceIdResponse(ResponseEntity<Object> responseEntity,
                                                    Class<?> clazz) throws JsonProcessingException {
-        HttpStatus status = responseEntity.getStatusCode();
+        HttpStatusCode status = responseEntity.getStatusCode();
 
         if (status.is2xxSuccessful()) {
             return objectMapper.convertValue(responseEntity.getBody(), clazz);
