@@ -33,7 +33,7 @@ public class KeyGenUtil {
         RSAKey rsaKey = KeyGenUtil.getRsaJwk();
         Map<String, List<JSONObject>> body = new LinkedHashMap<>();
         List<JSONObject> keyList = new ArrayList<>();
-        keyList.add(rsaKey.toJSONObject());
+        keyList.add((JSONObject) rsaKey.toJSONObject());
         body.put("keys", keyList);
         ObjectMapper objectMapper = new ObjectMapper();
         return objectMapper.writeValueAsString(body);
