@@ -1,47 +1,50 @@
 package uk.gov.hmcts.reform.cdapi.config;
 
+import io.swagger.v3.oas.annotations.enums.ParameterIn;
+import io.swagger.v3.oas.annotations.enums.SecuritySchemeIn;
+import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
+import io.swagger.v3.oas.annotations.security.SecurityScheme;
+import io.swagger.v3.oas.models.Operation;
+import io.swagger.v3.oas.models.media.StringSchema;
+import io.swagger.v3.oas.models.parameters.Parameter;
+import org.springdoc.core.customizers.OperationCustomizer;
+import org.springdoc.core.models.GroupedOpenApi;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.bind.annotation.RestController;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiKey;
-import springfox.documentation.service.SecurityScheme;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-import java.util.List;
-import java.util.Optional;
-
+import org.springframework.web.method.HandlerMethod;
 
 @Configuration
-@EnableSwagger2
+@SecurityScheme(name = "Authorization", type = SecuritySchemeType.HTTP, bearerFormat = "JWT", scheme = "bearer")
+@SecurityScheme(name = "ServiceAuthorization", type = SecuritySchemeType.APIKEY,
+    in = SecuritySchemeIn.HEADER, bearerFormat = "JWT", description = "ServiceAuthorization")
 public class SwaggerConfiguration {
 
     @Bean
-    public Docket api() {
-
-        return new Docket(DocumentationType.SWAGGER_2)
-            .useDefaultResponseMessages(false)
-            .genericModelSubstitutes(Optional.class)
-
-            .select()
-            .apis(RequestHandlerSelectors.withClassAnnotation(RestController.class))
-            .paths(PathSelectors.any())
-            .paths(PathSelectors.ant("/refdata/commondata/caseflags").negate())
-            .paths(PathSelectors.ant("/refdata/commondata/lov/categories").negate())
-            .build()
-            .securitySchemes(apiKeyList());
+    public GroupedOpenApi publicApi(OperationCustomizer customGlobalHeaders) {
+        return GroupedOpenApi.builder()
+            .group("rd-commondata-api")
+            .pathsToMatch("/**")
+            .build();
     }
 
-    private  List<SecurityScheme> apiKeyList() {
-        return
-            List.of(
-                new ApiKey("Authorization", "Authorization","header"),
-                new ApiKey("ServiceAuthorization", "ServiceAuthorization", "header")
-
-            );
+    @Bean
+    public OperationCustomizer customGlobalHeaders() {
+        return (Operation customOperation, HandlerMethod handlerMethod) -> {
+            Parameter serviceAuthorizationHeader = new Parameter()
+                .in(ParameterIn.HEADER.toString())
+                .schema(new StringSchema())
+                .name("ServiceAuthorization")
+                .description("Keyword `Bearer` followed by a service-to-service token for a whitelisted micro-service")
+                .required(true);
+            Parameter authorizationHeader = new Parameter()
+                .in(ParameterIn.HEADER.toString())
+                .schema(new StringSchema())
+                .name("Authorization")
+                .description("Authorization token")
+                .required(true);
+            customOperation.addParametersItem(authorizationHeader);
+            customOperation.addParametersItem(serviceAuthorizationHeader);
+            return customOperation;
+        };
     }
-
 }
