@@ -1,11 +1,12 @@
 package uk.gov.hmcts.reform.cdapi.controllers;
 
 
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
-import io.swagger.annotations.Authorization;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,8 +21,8 @@ import uk.gov.hmcts.reform.cdapi.exception.InvalidRequestException;
 import uk.gov.hmcts.reform.cdapi.service.CaseFlagService;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+import static uk.gov.hmcts.reform.cdapi.util.ValidationUtil.validateValueForYorNRequired;
 import static uk.gov.hmcts.reform.cdapi.util.ValidationUtil.validationFlagType;
-import static uk.gov.hmcts.reform.cdapi.util.ValidationUtil.validationWelshRequired;
 
 @RestController
 @Slf4j
@@ -33,52 +34,59 @@ public class CaseFlagApiController {
     @Autowired
     CaseFlagService caseFlagService;
 
-    @ApiOperation(
-        value = "CommonData API will be used to retrieve the list of case flags for a "
+    @Operation(
+        summary = "CommonData API will be used to retrieve the list of case flags for a "
             + "given service id.",
-        notes = "Any valid IDAM role should be able to access this API ",
-        authorizations = {
-            @Authorization(value = "ServiceAuthorization"),
-            @Authorization(value = "Authorization")
-        }
+        description = "Any valid IDAM role should be able to access this API ",
+        security =
+            {
+                @SecurityRequirement(name = "Authorization"),
+                @SecurityRequirement(name = "ServiceAuthorization")
+            }
     )
-    @ApiResponses({
-        @ApiResponse(
-            code = 200,
-            message = "Successfully retrieved list of Case Flag for the request provided",
-            response = CaseFlag.class
-        ),
-        @ApiResponse(
-            code = 400,
-            message = "Bad Request"
-        ),
-        @ApiResponse(
-            code = 401,
-            message = "Forbidden Error: Access denied"
-        ),
-        @ApiResponse(
-            code = 500,
-            message = "Internal Server Error"
-        )
-    })
+
+    @ApiResponse(
+        responseCode = "200",
+        description = "Successfully retrieved list of Case Flag for the request provided",
+        content = @Content(schema = @Schema(implementation = CaseFlag.class))
+    )
+    @ApiResponse(
+        responseCode = "400",
+        description = "Bad Request",
+        content = @Content
+    )
+    @ApiResponse(
+        responseCode = "401",
+        description = "Forbidden Error: Access denied",
+        content = @Content
+    )
+    @ApiResponse(
+        responseCode = "500",
+        description = "Internal Server Error",
+        content = @Content
+    )
     @GetMapping(
         produces = APPLICATION_JSON_VALUE,
-        path = {"/caseflags", "/caseflags/service-id={service-id}"}
+        path = {"/caseflags/service-id={service-id}"}
     )
     public ResponseEntity<CaseFlag> retrieveCaseFlagsByServiceId(
         @PathVariable(value = "service-id")
-        @ApiParam(name = "service-id",
-            value = "Any Valid String is allowed",
+        @Parameter(name = "service-id",
+            description = "Any Valid String is allowed",
             required = true)
-            String serviceId,
+        String serviceId,
         @RequestParam(value = "flag-type", required = false)
-        @ApiParam(name = "flag-type",
-            value = "Allowed Values are PARTY or CASE")
-            String flagType,
+        @Parameter(name = "flag-type",
+            description = "Allowed Values are PARTY or CASE")
+        String flagType,
         @RequestParam(value = "welsh-required", required = false)
-        @ApiParam(name = "welsh-required",
-            value = "Allowed Values are Y or N")
-            String welshRequired
+        @Parameter(name = "welsh-required",
+            description = "Allowed Values are Y or N")
+        String welshRequired,
+        @RequestParam(value = "available-external-flag", required = false)
+        @Parameter(name = "available-external-flag",
+            description = "Allowed Values are Y or N")
+        String availableExternalFlag
     ) {
         if (StringUtils.isEmpty(serviceId)) {
             throw new InvalidRequestException("service Id can not be null or empty");
@@ -87,10 +95,15 @@ public class CaseFlagApiController {
             validationFlagType(flagType.trim());
         }
         if (null != welshRequired) {
-            validationWelshRequired(welshRequired.trim());
+            validateValueForYorNRequired(welshRequired.trim());
+        }
+        if (null != availableExternalFlag) {
+            validateValueForYorNRequired(availableExternalFlag.trim());
         }
         log.info("Calling Service layer");
-        var caseFlag = caseFlagService.retrieveCaseFlagByServiceId(serviceId, flagType, welshRequired);
+        var caseFlag = caseFlagService.retrieveCaseFlagByServiceId(serviceId, flagType,
+                                                                   welshRequired, availableExternalFlag
+        );
         return ResponseEntity.ok().body(caseFlag);
     }
 
