@@ -1,9 +1,12 @@
 package uk.gov.hmcts.reform.cdapi.controllers;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.http.ResponseEntity.ok;
 
 /**
@@ -21,7 +24,17 @@ public class RootController {
      *
      * @return Welcome message from the service.
      */
-    @GetMapping("/")
+    @Operation(
+        summary = "Welcome"
+    )
+    @ApiResponse(
+        responseCode = "200",
+        description = "Welcome message"
+    )
+    @GetMapping(
+        path = "/",
+        produces = APPLICATION_JSON_VALUE
+    )
     public ResponseEntity<String> welcome() {
         return ok("Welcome to RD Common Data API");
     }
