@@ -1,6 +1,6 @@
 package uk.gov.hmcts.reform.cdapi.controllers.request;
 
-import io.swagger.annotations.ApiParam;
+import io.swagger.v3.oas.annotations.Parameter;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -12,21 +12,21 @@ import lombok.Setter;
 @AllArgsConstructor
 public class CategoryRequest {
 
-    @ApiParam(hidden = true)
+    @Parameter(hidden = true)
     String categoryId;
 
-    @ApiParam(name = "serviceId", value = "Any Valid String is allowed")
+    @Parameter(name = "serviceId", description = "Any Valid String is allowed")
     String serviceId;
 
-    @ApiParam(name = "parentCategory", value = "Any Valid String is allowed")
+    @Parameter(name = "parentCategory", description = "Any Valid String is allowed")
     String parentCategory;
 
-    @ApiParam(name = "parentKey", value = "Any Valid String is allowed")
+    @Parameter(name = "parentKey", description = "Any Valid String is allowed")
     String parentKey;
 
-    @ApiParam(name = "key", value = "Any Valid String is allowed")
+    @Parameter(name = "key", description = "Any Valid String is allowed")
     String key;
 
-    @ApiParam(name = "isChildRequired", value = "Any Valid String is allowed")
+    @Parameter(name = "isChildRequired", description = "Any Valid String is allowed")
     String isChildRequired;
 }
