@@ -1,10 +1,10 @@
 package uk.gov.hmcts.reform.cdapi.domain;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
 import lombok.Data;
 
 import java.io.Serializable;
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
 
 @Data
 @Embeddable
@@ -12,5 +12,7 @@ public class CategoryKey implements Serializable {
     @Column(name = "categorykey")
     private String categoryKey;
     private String key;
+    @Column(name = "serviceid")
+    private String serviceId;
 }
 
