@@ -1,11 +1,13 @@
 package uk.gov.hmcts.reform.cdapi.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 import lombok.Data;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
+import uk.gov.hmcts.reform.cdapi.domain.jsonfilter.IgnoreJsonFilter;
 
 @Data
 @Entity
@@ -17,4 +19,8 @@ public class ListOfValue {
     private String id;
     private String key;
     private String value;
+    @Column(name = "value_cy")
+    @JsonProperty("value_cy")
+    @JsonInclude(value = JsonInclude.Include.CUSTOM, valueFilter = IgnoreJsonFilter.class)
+    private String valueCy;
 }
