@@ -14,11 +14,12 @@ public interface CaseFlagRepository extends JpaRepository<CaseFlagDto, Long> {
 
     @Query(value = "with recursive flagdetailscategorypath"
         + "    AS ( "
-        + "    SELECT a.id, a.flag_code, a.value_en ,a.value_cy, a.category_id, a.value_en AS categorypath "
+        + "    SELECT a.id, a.flag_code, a.flag_code as native_flag_code, a.value_en ,a.value_cy, a.category_id, "
+        + "       a.value_en AS categorypath "
         + "       FROM flag_details a "
         + "        WHERE a.category_id = 0 "
         + "       UNION ALL "
-        + "       SELECT b.id, b.flag_code, b.value_en, b.value_cy, b.category_id, "
+        + "       SELECT b.id, b.flag_code, b.flag_code as native_flag_code, b.value_en, b.value_cy, b.category_id, "
         + "       concat(c.categorypath,'/',b.value_en) AS categorypath "
         + "         FROM flag_details b "
         + "       INNER JOIN flagdetailscategorypath c "
@@ -26,14 +27,16 @@ public interface CaseFlagRepository extends JpaRepository<CaseFlagDto, Long> {
         + "       ) "
         + "     , flagserviceconsolidated "
         + "    AS ( "
-        + "       SELECT j.flag_code, j.hearing_relevant, j.request_reason "
+        + "       SELECT j.flag_code, j.flag_code as native_flag_code, j.hearing_relevant, "
+        + "       j.request_reason, j.default_status, j.available_externally "
         + "         FROM flag_service j "
         + "        WHERE j.service_id = 'XXXX' "
         + "          AND j.flag_code NOT IN (SELECT k.flag_code "
         + "                                    FROM flag_service k "
         + "                                   WHERE upper(k.service_id) = :serviceId) "
         + "       UNION "
-        + "       SELECT l.flag_code, l.hearing_relevant, l.request_reason "
+        + "       SELECT l.flag_code, l.flag_code as native_flag_code, l.hearing_relevant, "
+        + "       l.request_reason, l.default_status, l.available_externally "
         + "         FROM flag_service l "
         + "        WHERE upper(l.service_id) = :serviceId "
         + "       ) "
@@ -46,35 +49,44 @@ public interface CaseFlagRepository extends JpaRepository<CaseFlagDto, Long> {
         + "       ) "
         + "     , relevantcategories "
         + "    AS ( "
-        + "       SELECT s.id, s.flag_code, s.value_en, s.value_cy, s.category_id, s.value_en AS categorypath "
+        + "       SELECT s.id, s.flag_code, s.flag_code as native_flag_code, s.value_en, "
+        + "       s.value_cy, s.category_id, s.value_en AS categorypath "
         + "         FROM flag_details s "
         + "        WHERE s.id IN (SELECT t.category_id "
         + "                         FROM distinctcategoryid t) "
         + "       UNION ALL "
-        + "       SELECT u.id, u.flag_code, u.value_en, u.value_cy, u.category_id, u.value_en AS categorypath "
+        + "       SELECT u.id, u.flag_code, u.flag_code as native_flag_code, u.value_en, u.value_cy, "
+        + "       u.category_id, u.value_en AS categorypath "
         + "         FROM flag_details u "
         + "       INNER JOIN relevantcategories v "
         + "               ON u.id = v.category_id "
         + "       ) "
-        + "SELECT h.id, h.flag_code, h.value_en, h.value_cy, h.category_id, f.categorypath, "
-        + "       h.hearing_relevant, h.request_reason, h.isparent "
+        + "SELECT h.id, h.flag_code, h.native_flag_code, h.value_en, h.value_cy, h.category_id, f.categorypath, "
+        + "       h.hearing_relevant, h.request_reason, h.default_status, h.available_externally, h.isparent "
         + "  FROM flagdetailscategorypath f "
         + "INNER JOIN "
         + "( "
-        + " SELECT d.id, d.flag_code, d.value_en, d.value_cy, d.category_id, d.categorypath, "
-        + "        e.hearing_relevant, e.request_reason, 'FALSE' AS isparent "
+        + " SELECT d.id, d.flag_code, d.flag_code as native_flag_code, d.value_en, d.value_cy, "
+        + " d.category_id, d.categorypath, "
+        + "        e.hearing_relevant, e.request_reason, e.default_status, e.available_externally, "
+        + "        'FALSE' AS isparent "
         + "   FROM flagdetailscategorypath d "
         + " INNER JOIN flagserviceconsolidated e "
         + "         ON d.flag_code = e.flag_code "
+        + "         WHERE d.id not in (SELECT category_id from flagdetailscategorypath)"
         + " UNION "
-        + " SELECT p.id, 'CATGRY' AS flag_code, p.value_en, p.value_cy, p.category_id, p.categorypath, "
-        + "        'FALSE' AS hearing_relevant, 'FALSE' AS request_reason, 'TRUE' AS isparent "
+        + " SELECT p.id, 'CATGRY' AS flag_code, p.native_flag_code, p.value_en, p.value_cy, "
+        + " p.category_id, p.categorypath, "
+        + "        'FALSE' AS hearing_relevant, 'FALSE' AS request_reason, 'Active' as default_status, "
+        + "'FALSE' as available_externally, 'TRUE' AS isparent "
         + "   FROM relevantcategories p "
         + ") h "
         + "ON f.id = h.category_id "
         + "UNION "
-        + "SELECT q.id, 'CATGRY' AS flag_code, q.value_en, q.value_cy, q.category_id, '' AS categorypath, "
-        + "       'FALSE' AS hearing_relevant, 'FALSE' AS request_reason, 'TRUE' AS isparent "
+        + "SELECT q.id, 'CATGRY' AS flag_code, q.flag_code as native_flag_code, q.value_en, q.value_cy, "
+        + "q.category_id, '' AS categorypath, "
+        + "       'FALSE' AS hearing_relevant, 'FALSE' AS request_reason, 'Active' as default_status, "
+        + "'FALSE' as available_externally, 'TRUE' AS isparent "
         + "  FROM flag_details q "
         + " INNER JOIN relevantcategories r "
         + "         ON r.id = q.id "
