@@ -2,6 +2,7 @@ package uk.gov.hmcts.reform.cdapi.service.impl;
 
 import com.launchdarkly.sdk.LDUser;
 import com.launchdarkly.sdk.server.LDClient;
+import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -9,7 +10,6 @@ import uk.gov.hmcts.reform.cdapi.service.FeatureToggleService;
 
 import java.util.HashMap;
 import java.util.Map;
-import javax.annotation.PostConstruct;
 
 @Service
 public class FeatureToggleServiceImpl implements FeatureToggleService {
