@@ -1,101 +1,294 @@
 package uk.gov.hmcts.reform.cdapi.controllers;
 
-import org.junit.jupiter.api.BeforeEach;
+import org.jetbrains.annotations.NotNull;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.boot.test.mock.mockito.SpyBean;
+import org.springframework.http.MediaType;
+import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.ResultActions;
 import uk.gov.hmcts.reform.cdapi.controllers.request.CategoryRequest;
-import uk.gov.hmcts.reform.cdapi.controllers.response.Categories;
 import uk.gov.hmcts.reform.cdapi.controllers.response.Category;
-import uk.gov.hmcts.reform.cdapi.exception.InvalidRequestException;
-import uk.gov.hmcts.reform.cdapi.service.CrdService;
+import uk.gov.hmcts.reform.cdapi.exception.ResourceNotFoundException;
+import uk.gov.hmcts.reform.cdapi.exception.handler.GlobalExceptionHandler;
+import uk.gov.hmcts.reform.cdapi.service.impl.CrdServiceImpl;
 
-import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.apache.commons.lang3.RandomStringUtils.randomAlphabetic;
+import static org.apache.commons.lang3.RandomUtils.nextLong;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.nullValue;
+import static org.hamcrest.Matchers.hasSize;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.reform.cdapi.helper.CrdTestSupport.buildCategoryRequest;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.then;
+import static org.mockito.BDDMockito.willThrow;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+@WebMvcTest
+@WithMockUser
+@ContextConfiguration(classes = CrdApiController.class)
 class CrdApiControllerTest {
 
-    @Mock
-    CrdService crdService;
+    @MockBean
+    CrdServiceImpl crdService;
 
-    @InjectMocks
-    CrdApiController crdApiController;
+    @SpyBean
+    private GlobalExceptionHandler globalExceptionHandler;
 
-    @BeforeEach
-    public void setUp() {
-        MockitoAnnotations.openMocks(this);
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Test
+    @DisplayName("Positive scenario -Should return 200 with list of categories for parent category without child nodes")
+    void should_Return_200_With_Categories_For_CategoryId() throws Exception {
+
+        final String categoryId = "HearingChannel";
+        final boolean isChildRequired = false;
+        final List<Category> categoryList = createCategoryList(categoryId, isChildRequired);
+
+        // given
+        given(crdService.retrieveListOfValuesByCategory(any(CategoryRequest.class))).willReturn(categoryList);
+
+        //when
+        final ResultActions resultActions =
+            mockMvc.perform(get("/refdata/commondata/lov/categories/{categoryId}", categoryId)
+                                .accept(MediaType.APPLICATION_JSON_VALUE))
+                .andDo(print())
+                .andExpect(status().isOk());
+
+        //then
+        assertParentContent(resultActions, categoryList);
+        final List<Category> childNodes = categoryList.get(0).getChildNodes();
+        assertThat(childNodes).isNull();
+        then(crdService).should().retrieveListOfValuesByCategory(any(CategoryRequest.class));
     }
 
     @Test
-    void testWithValidCategoryId_ShouldReturnStatusCode200() {
-        CategoryRequest request = buildCategoryRequest(null, null, null,
-                                                       null, null, "N");
-        List<Category> categoryList = List.of(Category.builder().categoryKey("HearingChannel")
-                                                  .childNodes(Collections.emptyList()).build());
+    @DisplayName("Positive scenario - Should return 200 with list of categories for parent category with child nodes")
+    void should_Return_200_With_Categories_With_ChildNodes_For_Parent_CategoryId() throws Exception {
 
-        when(crdService.retrieveListOfValuesByCategory(any(CategoryRequest.class))).thenReturn(categoryList);
+        final String parentCategoryId = "HearingChannel";
+        final List<Category> categoryList = createCategoryList(parentCategoryId, true);
 
-        ResponseEntity<Categories> result = crdApiController.retrieveListOfValuesByCategoryId(
-            java.util.Optional.of("HearingChannel"), request);
-        assertNotNull(result.getBody());
-        assertFalse(result.getBody().getListOfCategory().isEmpty());
-        assertTrue(result.getBody().getListOfCategory().get(0).getChildNodes().isEmpty());
-        assertEquals(HttpStatus.OK, result.getStatusCode());
-        assertEquals("HearingChannel", request.getCategoryId());
-        verify(crdService, times(1)).retrieveListOfValuesByCategory(request);
+        // given
+        given(crdService.retrieveListOfValuesByCategory(any(CategoryRequest.class))).willReturn(categoryList);
+
+        //when
+        final ResultActions resultActions =
+            mockMvc.perform(get("/refdata/commondata/lov/categories/{categoryId}", parentCategoryId)
+                                .queryParam("serviceId", randomAlphabetic(5))
+                                .queryParam("key", randomAlphabetic(5))
+                                .queryParam("isChildRequired", "Y")
+                                .accept(MediaType.APPLICATION_JSON_VALUE))
+                .andDo(print())
+                .andExpect(status().isOk());
+
+        //then
+        assertParentContent(resultActions, categoryList);
+        final List<Category> childNodes = categoryList.get(0).getChildNodes();
+        assertThat(childNodes).isNotNull();
+        assertChildContent(resultActions, categoryList);
+        then(crdService).should().retrieveListOfValuesByCategory(any(CategoryRequest.class));
     }
 
     @Test
-     void whenIdIsNull_thenExceptionIsThrown() {
-        CategoryRequest request = buildCategoryRequest("HearingChannel", null, null,
-                                                       null, null, "N");
+    @DisplayName("Positive scenario -Should return 200 with list of categories for parent category without child nodes")
+    void should_Return_200_With_Categories_Without_ChildNodes_For_CategoryId() throws Exception {
 
-        Optional<String> empty = Optional.empty();
-        assertThrows(InvalidRequestException.class, () -> crdApiController.retrieveListOfValuesByCategoryId(
-            empty, request));
+        final String categoryId = "HearingChannel";
+        final List<Category> categoryList = createCategoryList(categoryId, false);
 
+        // given
+        given(crdService.retrieveListOfValuesByCategory(any(CategoryRequest.class))).willReturn(categoryList);
+
+        //when
+        final ResultActions resultActions =
+            mockMvc.perform(get("/refdata/commondata/lov/categories/{categoryId}", categoryId)
+                                .queryParam("serviceId", randomAlphabetic(5))
+                                .queryParam("key", randomAlphabetic(5))
+                                .queryParam("isChildRequired", "N")
+                                .accept(MediaType.APPLICATION_JSON_VALUE))
+                .andDo(print())
+                .andExpect(status().isOk());
+
+        //then
+        assertParentContent(resultActions, categoryList);
+        final List<Category> childNodes = categoryList.get(0).getChildNodes();
+        assertThat(childNodes).isNull();
+        then(crdService).should().retrieveListOfValuesByCategory(any(CategoryRequest.class));
     }
 
     @Test
-    void testWithAllValidParamValues_ShouldReturnStatusCode200() {
-        CategoryRequest request = buildCategoryRequest(null, "BBA3", "1",
-                                                       "1", "5", "y");
-        when(crdService.retrieveListOfValuesByCategory(any(CategoryRequest.class))).thenReturn(Collections.emptyList());
+    @DisplayName("Positive scenario - Should return 200 with list of categories for sub-category without child nodes")
+    void should_Return_200_With_Categories_And_Without_ChildNodes_For_SubCategoryId() throws Exception {
 
-        Optional<String> key = java.util.Optional.of("HearingChannel");
-        ResponseEntity<Categories> result = crdApiController.retrieveListOfValuesByCategoryId(key, request);
-        assertNotNull(result.getBody());
-        assertEquals("[]", result.getBody().getListOfCategory().toString());
-        assertTrue(result.getBody().getListOfCategory().isEmpty());
-        assertEquals(HttpStatus.OK, result.getStatusCode());
-        assertEquals(request.getCategoryId(), key.get());
-        verify(crdService, times(1)).retrieveListOfValuesByCategory(request);
+        final String subCategoryId = "SubHearingChannel";
+        final List<Category> categoryList = createCategoryList(subCategoryId, false);
+
+        // given
+        given(crdService.retrieveListOfValuesByCategory(any(CategoryRequest.class))).willReturn(categoryList);
+
+        //when
+        final ResultActions resultActions =
+            mockMvc.perform(get("/refdata/commondata/lov/categories/{categoryId}", subCategoryId)
+                                .queryParam("serviceId", randomAlphabetic(5))
+                                .queryParam("parentCategory", randomAlphabetic(5))
+                                .queryParam("parentKey", randomAlphabetic(5))
+                                .queryParam("key", randomAlphabetic(5))
+                                .queryParam("isChildRequired", "N")
+                                .accept(MediaType.APPLICATION_JSON_VALUE))
+                .andDo(print())
+                .andExpect(status().isOk());
+
+        //then
+        assertParentContent(resultActions, categoryList);
+        final List<Category> childNodes = categoryList.get(0).getChildNodes();
+        assertThat(childNodes).isNull();
+        then(crdService).should().retrieveListOfValuesByCategory(any(CategoryRequest.class));
+    }
+
+
+    @Test
+    @DisplayName("Negative scenario - Should return 404 when categories not found")
+    void should_Return_404_When_Categories_Not_Found() throws Exception {
+
+        //given
+        willThrow(ResourceNotFoundException.class)
+            .given(crdService).retrieveListOfValuesByCategory(any(CategoryRequest.class));
+
+        //when
+        mockMvc.perform(get("/refdata/commondata/lov/categories/{categoryId}", randomAlphabetic(5))
+                            .accept(MediaType.APPLICATION_JSON_VALUE))
+            .andDo(print())
+
+            //then
+            .andExpect(status().isNotFound())
+            .andExpect(jsonPath("$.errorCode", is(404)))
+            .andExpect(jsonPath("$.status", is("Not Found")))
+            .andExpect(jsonPath("$.errorMessage", is("4 : Resource not found")));
+
+        then(crdService).should().retrieveListOfValuesByCategory(any(CategoryRequest.class));
     }
 
     @Test
-    void testWithOnlyCategoryId() {
-        CategoryRequest request = mock(CategoryRequest.class);
-        Optional<String> key = java.util.Optional.of("HearingChannel");
-        ResponseEntity<?> result = crdApiController.retrieveListOfValuesByCategoryId(key, request);
-        assertNotNull(result);
-        assertEquals(HttpStatus.OK, result.getStatusCode());
-        verify(request, times(1)).setCategoryId(any());
-        verify(crdService, times(1)).retrieveListOfValuesByCategory(request);
+    @DisplayName("Negative scenario - Should return Not Found when categoryId not present")
+    void should_Return_NotFound_When_CategoryId_Is_Not_Present() throws Exception {
+
+        //when
+        mockMvc
+            .perform(get("/refdata/commondata/lov/categories/{categoryId}", "")
+                         .accept(MediaType.APPLICATION_JSON_VALUE))
+            .andDo(print())
+            //then
+            .andExpect(status().isNotFound());
+
+        then(crdService).shouldHaveNoInteractions();
     }
+
+    private static void assertParentContent(final ResultActions resultActions,
+                                            final List<Category> categoryList)
+        throws Exception {
+        final Category parentCategoryListDetails = categoryList.get(0);
+
+        resultActions
+            .andExpect(jsonPath("$.list_of_values", hasSize(1)))
+
+            .andExpect(jsonPath("$.list_of_values[0].category_key", is(parentCategoryListDetails.getCategoryKey())))
+            .andExpect(jsonPath("$.list_of_values[0].key", is(parentCategoryListDetails.getKey())))
+            .andExpect(jsonPath("$.list_of_values[0].value_en", is(parentCategoryListDetails.getValueEn())))
+            .andExpect(jsonPath("$.list_of_values[0].value_cy", is(parentCategoryListDetails.getValueCy())))
+            .andExpect(jsonPath("$.list_of_values[0].hint_text_en", is(parentCategoryListDetails.getHintTextEn())))
+            .andExpect(jsonPath("$.list_of_values[0].hint_text_cy", is(parentCategoryListDetails.getHintTextCy())))
+            .andExpect(jsonPath("$.list_of_values[0].lov_order", is(parentCategoryListDetails.getLovOrder())))
+            .andExpect(jsonPath(
+                "$.list_of_values[0].parent_category",
+                is(parentCategoryListDetails.getParentCategory())
+            ))
+            .andExpect(jsonPath("$.list_of_values[0].parent_key", is(parentCategoryListDetails.getParentKey())))
+            .andExpect(jsonPath("$.list_of_values[0].active_flag", is(parentCategoryListDetails.getActiveFlag())));
+    }
+
+    private static void assertChildContent(ResultActions resultActions, List<Category> categoryList)
+        throws Exception {
+        final Category childCategoryListDetails = categoryList.get(0).getChildNodes().get(0);
+
+        resultActions
+            .andExpect(jsonPath("$.list_of_values[0].child_nodes", hasSize(1)))
+            .andExpect(jsonPath(
+                "$.list_of_values[0].child_nodes[0].category_key",
+                is(childCategoryListDetails.getCategoryKey())
+            ))
+            .andExpect(jsonPath("$.list_of_values[0].child_nodes[0].key", is(childCategoryListDetails.getKey())))
+            .andExpect(jsonPath(
+                "$.list_of_values[0].child_nodes[0].value_en",
+                is(childCategoryListDetails.getValueEn())
+            ))
+            .andExpect(jsonPath(
+                "$.list_of_values[0].child_nodes[0].value_cy",
+                is(childCategoryListDetails.getValueCy())
+            ))
+            .andExpect(jsonPath(
+                "$.list_of_values[0].child_nodes[0].hint_text_en",
+                is(childCategoryListDetails.getHintTextEn())
+            ))
+            .andExpect(jsonPath(
+                "$.list_of_values[0].child_nodes[0].hint_text_cy",
+                is(childCategoryListDetails.getHintTextCy())
+            ))
+            .andExpect(jsonPath(
+                "$.list_of_values[0].child_nodes[0].lov_order",
+                is(childCategoryListDetails.getLovOrder())
+            ))
+            .andExpect(jsonPath(
+                "$.list_of_values[0].child_nodes[0].parent_category",
+                is(childCategoryListDetails.getParentCategory())
+            ))
+            .andExpect(jsonPath(
+                "$.list_of_values[0].child_nodes[0].parent_key",
+                is(childCategoryListDetails.getParentKey())
+            ))
+            .andExpect(jsonPath(
+                "$.list_of_values[0].child_nodes[0].active_flag",
+                is(childCategoryListDetails.getActiveFlag())
+            ))
+            .andExpect(jsonPath("$.list_of_values[0].child_nodes[0].child_nodes", is(nullValue())));
+    }
+
+    @NotNull
+    private static List<Category> createCategoryList(final String categoryId,
+                                                     final boolean isChildRequired) {
+
+        final Category.CategoryBuilder categoryBuilder =
+            Category
+                .builder()
+                .categoryKey(categoryId)
+                .key(randomAlphabetic(5))
+                .valueEn(randomAlphabetic(5))
+                .valueCy(randomAlphabetic(5))
+                .hintTextEn(randomAlphabetic(5))
+                .hintTextCy(randomAlphabetic(5))
+                .lovOrder(nextLong())
+                .parentKey(randomAlphabetic(5))
+                .activeFlag(randomAlphabetic(5))
+                .parentCategory(randomAlphabetic(5));
+
+        if (isChildRequired) {
+            categoryBuilder.childNodes(List.of(categoryBuilder.categoryKey("Sub".concat(categoryId)).build()));
+        }
+
+        return List.of(categoryBuilder.build());
+    }
+
+
 }
