@@ -45,26 +45,26 @@ class JwtGrantedAuthoritiesConverterTest {
 
     @Test
     void test_shouldReturnEmptyAuthoritiesWhenClaimNotAvailable() {
-        when(jwtMock.containsClaim(anyString())).thenReturn(false);
+        when(jwtMock.hasClaim(anyString())).thenReturn(false);
 
         Collection<GrantedAuthority> authorities = converter.convert(jwtMock);
 
         assertNotNull(authorities);
         assertEquals(0, authorities.size());
-        verify(jwtMock, times(1)).containsClaim(anyString());
+        verify(jwtMock, times(1)).hasClaim(anyString());
         verify(idamRepositoryMock, times(0)).getUserInfo(anyString());
     }
 
     @Test
     void test_shouldReturnEmptyAuthoritiesWhenClaimValueNotEquals() {
-        when(jwtMock.containsClaim(anyString())).thenReturn(true);
+        when(jwtMock.hasClaim(anyString())).thenReturn(true);
         when(jwtMock.getClaim(anyString())).thenReturn("Test");
 
         Collection<GrantedAuthority> authorities = converter.convert(jwtMock);
 
         assertNotNull(authorities);
         assertEquals(0, authorities.size());
-        verify(jwtMock, times(1)).containsClaim(anyString());
+        verify(jwtMock, times(1)).hasClaim(anyString());
         verify(jwtMock, times(1)).getClaim(anyString());
         verify(idamRepositoryMock, times(0)).getUserInfo(anyString());
     }
@@ -73,7 +73,7 @@ class JwtGrantedAuthoritiesConverterTest {
     void test_shouldReturnEmptyAuthoritiesWhenIdamReturnsNoUsers() {
         List<String> roles = new ArrayList<>();
 
-        when(jwtMock.containsClaim(anyString())).thenReturn(true);
+        when(jwtMock.hasClaim(anyString())).thenReturn(true);
         when(jwtMock.getClaim(anyString())).thenReturn("access_token");
         when(jwtMock.getTokenValue()).thenReturn("access_token");
         when(userInfoMock.getRoles()).thenReturn(roles);
@@ -83,7 +83,7 @@ class JwtGrantedAuthoritiesConverterTest {
 
         assertNotNull(authorities);
         assertEquals(0, authorities.size());
-        verify(jwtMock, times(1)).containsClaim(anyString());
+        verify(jwtMock, times(1)).hasClaim(anyString());
         verify(jwtMock, times(1)).getClaim(anyString());
         verify(jwtMock, times(1)).getTokenValue();
         verify(userInfoMock, times(1)).getRoles();
@@ -95,7 +95,7 @@ class JwtGrantedAuthoritiesConverterTest {
     void test_shouldReturnEmptyAuthoritiesWhenIdamReturnsUsers() {
         List<String> roles = new ArrayList<>();
         roles.add("crd-admin");
-        when(jwtMock.containsClaim(anyString())).thenReturn(true);
+        when(jwtMock.hasClaim(anyString())).thenReturn(true);
         when(jwtMock.getClaim(anyString())).thenReturn("access_token");
         when(jwtMock.getTokenValue()).thenReturn("access_token");
         when(userInfoMock.getRoles()).thenReturn(roles);
@@ -105,7 +105,7 @@ class JwtGrantedAuthoritiesConverterTest {
 
         assertNotNull(authorities);
         assertEquals(1, authorities.size());
-        verify(jwtMock, times(1)).containsClaim(anyString());
+        verify(jwtMock, times(1)).hasClaim(anyString());
         verify(jwtMock, times(1)).getClaim(anyString());
         verify(jwtMock, times(1)).getTokenValue();
         verify(userInfoMock, times(2)).getRoles();
@@ -116,7 +116,7 @@ class JwtGrantedAuthoritiesConverterTest {
     @Test
     void test_shouldReturnEmptyAuthoritiesWhenEmptyRoles() {
         List<String> roles = new ArrayList<>();
-        when(jwtMock.containsClaim(anyString())).thenReturn(true);
+        when(jwtMock.hasClaim(anyString())).thenReturn(true);
         when(jwtMock.getClaim(anyString())).thenReturn("access_token");
         when(jwtMock.getTokenValue()).thenReturn("access_token");
         when(userInfoMock.getRoles()).thenReturn(roles);
@@ -126,7 +126,7 @@ class JwtGrantedAuthoritiesConverterTest {
 
         assertNotNull(authorities);
         assertEquals(0, authorities.size());
-        verify(jwtMock, times(1)).containsClaim(anyString());
+        verify(jwtMock, times(1)).hasClaim(anyString());
         verify(jwtMock, times(1)).getClaim(anyString());
         verify(jwtMock, times(1)).getTokenValue();
         verify(userInfoMock, times(1)).getRoles();
