@@ -1,8 +1,11 @@
 package uk.gov.hmcts.reform.cdapi.service.impl;
 
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -23,12 +26,14 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.cdapi.controllers.constant.Constant.FLAG_PF0015;
 import static uk.gov.hmcts.reform.cdapi.controllers.constant.Constant.FLAG_RA0042;
+import static uk.gov.hmcts.reform.cdapi.service.impl.CaseFlagServiceImpl.IGNORE_JSON;
 
 @ExtendWith(MockitoExtension.class)
 class CaseFlagServiceImplTest {
@@ -49,33 +54,25 @@ class CaseFlagServiceImplTest {
         );
     }
 
-    @Test
-    void testGetCaseFlag_ByServiceId_Returns200() {
+    @ParameterizedTest
+    @CsvSource({
+        "XXXX,,N,N",
+        "XXXX,PARTY,null,null",
+        "XXXX,CASE,n,n"
+    })
+    void testGetCaseFlag_ByServiceId_Returns200(String serviceId,String flagType,String welshRequired,
+                                                String availableExternalFlag) {
         when(caseFlagRepository.findAll(anyString())).thenReturn(getCaseFlagDtoList());
-        var caseFlag = caseFlagService.retrieveCaseFlagByServiceId("XXXX", "", "N");
+        var caseFlag = caseFlagService
+            .retrieveCaseFlagByServiceId(serviceId, flagType, welshRequired, availableExternalFlag);
         assertNotNull(caseFlag);
         assertEquals(1, caseFlag.getFlags().size());
-        assertEquals(2, caseFlag.getFlags().get(0).getFlagDetails().size());
-        verify(caseFlagRepository, times(1)).findAll(anyString());
-    }
+        if (!(flagType == null)) {
+            assertEquals(flagType, caseFlag.getFlags().get(0).getFlagDetails().get(0).getName());
+        } else {
+            assertEquals(2, caseFlag.getFlags().get(0).getFlagDetails().size());
+        }
 
-    @Test
-    void testGetCaseFlag_ByServiceIdAndFlagTypeParty() {
-        when(caseFlagRepository.findAll(anyString())).thenReturn(getCaseFlagDtoList());
-        var caseFlag = caseFlagService.retrieveCaseFlagByServiceId("XXXX", "PARTY", null);
-        assertNotNull(caseFlag);
-        assertEquals(1, caseFlag.getFlags().get(0).getFlagDetails().size());
-        assertEquals("PARTY", caseFlag.getFlags().get(0).getFlagDetails().get(0).getName());
-        verify(caseFlagRepository, times(1)).findAll(anyString());
-    }
-
-    @Test
-    void testGetCaseFlag_ByServiceIdAndFlagTypeCase() {
-        when(caseFlagRepository.findAll(anyString())).thenReturn(getCaseFlagDtoList());
-        var caseFlag = caseFlagService.retrieveCaseFlagByServiceId("XXXX", "CASE", "n");
-        assertNotNull(caseFlag);
-        assertEquals(1, caseFlag.getFlags().get(0).getFlagDetails().size());
-        assertEquals("CASE", caseFlag.getFlags().get(0).getFlagDetails().get(0).getName());
         verify(caseFlagRepository, times(1)).findAll(anyString());
     }
 
@@ -83,29 +80,26 @@ class CaseFlagServiceImplTest {
     void testGetCaseFlag_ByServiceIdAndWelshRequiredasY() {
         when(caseFlagRepository.findAll(anyString())).thenReturn(getCaseFlagDtoList());
         assertThrows(ResourceNotFoundException.class, () ->
-            caseFlagService.retrieveCaseFlagByServiceId("XXXX", "CASE", "y"));
+            caseFlagService.retrieveCaseFlagByServiceId("XXXX", "TEST", "y", "n"));
     }
 
-    @Test
-    void testGetCaseFlag_WhenLanguageInterpreterFlag_return200() {
-        when(caseFlagRepository.findAll(anyString())).thenReturn(getCaseFlagDtoListWithLanguageInterpreter());
-        when(listOfVenueRepository.findListOfValues(anyString())).thenReturn(getListOfValuesForLanguageInterPreter());
-        var caseFlag = caseFlagService.retrieveCaseFlagByServiceId("XXXX", "PARTY", "");
+    @ParameterizedTest
+    @CsvSource({
+        "XXXX,PARTY,N,false",
+        "XXXX,PARTY,Y,true",
+        "XXXX,PARTY,N,false"
+    })
+    void testGetCaseFlags(String serviceId,String flagType,String welshRequired,
+                                                               boolean flag) {
+        when(caseFlagRepository.findAll(anyString()))
+            .thenReturn(getCaseFlagDtoListWithLanguageInterpreter());
+        when(listOfVenueRepository.findListOfValues(anyString()))
+            .thenReturn(getListOfValuesForLanguageInterPreter(flag));
+        var caseFlag = caseFlagService.retrieveCaseFlagByServiceId(serviceId, flagType, welshRequired, "");
         assertNotNull(caseFlag);
         verify(caseFlagRepository, times(1)).findAll(anyString());
         verify(listOfVenueRepository, times(1)).findListOfValues(anyString());
-        verifyListOfValuesResponse(caseFlag);
-    }
-
-    @Test
-    void testGetCaseFlag_WhenSignLanguageFlag_return200() {
-        when(caseFlagRepository.findAll(anyString())).thenReturn(getCaseFlagDtoListWithSignLanguage());
-        when(listOfVenueRepository.findListOfValues(anyString())).thenReturn(getListOfValuesForSignLanguage());
-        var caseFlag = caseFlagService.retrieveCaseFlagByServiceId("XXXX", "PARTY", "");
-        assertNotNull(caseFlag);
-        verify(caseFlagRepository, times(1)).findAll(anyString());
-        verify(listOfVenueRepository, times(1)).findListOfValues(anyString());
-        verifyListOfValuesResponse(caseFlag);
+        verifyListOfValuesResponse(caseFlag, flag);
     }
 
     @Test
@@ -113,8 +107,103 @@ class CaseFlagServiceImplTest {
         when(caseFlagRepository.findAll(anyString())).thenReturn(getCaseFlagDtoList());
         assertThrows(
             ResourceNotFoundException.class,
-            () -> caseFlagService.retrieveCaseFlagByServiceId("XXXX", "Hello", "")
+            () -> caseFlagService.retrieveCaseFlagByServiceId("XXXX", "Hello", "", "")
         );
+    }
+
+    @ParameterizedTest
+    @CsvSource({
+        "XXXX,Y,N",
+        "XXXX,N,N",
+        "XXXX,N,Y",
+        "XXXX,Y,Y",
+        "XXXX,N,N"
+    })
+    void testGetCaseFlag_ByServiceIWithWelshRequired200(String serviceId, String welshRequired,
+                                                                   String availableExternalFlag) {
+        when(caseFlagRepository.findAll(anyString())).thenReturn(getCaseFlagDtoList());
+        var caseFlag = caseFlagService.retrieveCaseFlagByServiceId(serviceId, "", welshRequired,
+                                                                   availableExternalFlag);
+        assertNotNull(caseFlag);
+        assertEquals(1, caseFlag.getFlags().size());
+        assertEquals(2, caseFlag.getFlags().get(0).getFlagDetails().size());
+        verify(caseFlagRepository, times(1)).findAll(anyString());
+        caseFlag.getFlags().forEach(caseFlagObj -> {
+            for (FlagDetail flagDetail : caseFlagObj.getFlagDetails()) {
+                if (("N").equals(welshRequired)) {
+                    assertEquals(IGNORE_JSON, flagDetail.getNameCy());
+                }
+                assertNotNull(flagDetail.getDefaultStatus());
+                assertNotNull(flagDetail.getExternallyAvailable());
+                assertTrue(flagDetail.getExternallyAvailable());
+            }
+        });
+    }
+
+    @Test
+    void testGetCaseFlag_ByServiceIWithFlagDetailsNull_Returns200() {
+        when(caseFlagRepository.findAll(anyString())).thenReturn(getEmptyCaseFlagDtoList(getCaseFlagDtoList()));
+        when(listOfVenueRepository.findListOfValues(anyString()))
+            .thenReturn(getListOfValuesForLanguageInterPreter(false));
+
+        var caseFlag = caseFlagService.retrieveCaseFlagByServiceId("XXXX", "", "N", "Y");
+        assertNotNull(caseFlag);
+        assertEquals(1, caseFlag.getFlags().size());
+        assertEquals(2, caseFlag.getFlags().get(0).getFlagDetails().size());
+        verify(caseFlagRepository, times(1)).findAll(anyString());
+        caseFlag.getFlags().forEach(caseFlagObj -> {
+            for (FlagDetail flagDetail : caseFlagObj.getFlagDetails()) {
+                assertNotNull(flagDetail.getDefaultStatus());
+                assertNotNull(flagDetail.getExternallyAvailable());
+            }
+        });
+    }
+
+    @Test
+    @DisplayName("Positive scenario -Should return 200 with Welsh-Required=N and available-externally=N")
+    void testGetCaseFlag_ByServiceIWithWelshRequiredIsNandAvailableExternallyIsN_Returns200() {
+        when(caseFlagRepository.findAll(anyString())).thenReturn(getCaseFlagDtoList());
+        var caseFlag = caseFlagService.retrieveCaseFlagByServiceId("XXXX", "", "N", "N");
+        assertNotNull(caseFlag);
+        assertEquals(1, caseFlag.getFlags().size());
+        assertEquals(2, caseFlag.getFlags().get(0).getFlagDetails().size());
+        verify(caseFlagRepository, times(1)).findAll(anyString());
+        caseFlag.getFlags().forEach(caseFlagObj -> {
+            for (FlagDetail flagDetail : caseFlagObj.getFlagDetails()) {
+                assertEquals(IGNORE_JSON, flagDetail.getNameCy());
+                assertNotNull(flagDetail.getDefaultStatus());
+                assertNotNull(flagDetail.getExternallyAvailable());
+            }
+        });
+    }
+
+    List<CaseFlagDto> getEmptyCaseFlagDtoList(List<CaseFlagDto> caseFlagDtoList) {
+        var caseFlagDto6 = new CaseFlagDto();
+        caseFlagDto6.setFlagCode("CATEGORY");
+        caseFlagDto6.setCategoryId(4);
+        caseFlagDto6.setCategoryPath("CASE");
+        caseFlagDto6.setId(5);
+        caseFlagDto6.setHearingRelevant(true);
+        caseFlagDto6.setRequestReason(false);
+        caseFlagDto6.setValueEn("COMPLEX CASE");
+        caseFlagDto6.setValueCy("");
+        caseFlagDto6.setIsParent(false);
+        caseFlagDto6.setExternallyAvailable(false);
+        caseFlagDto6.setDefaultStatus("Requested");
+        var caseFlagDto3 = new CaseFlagDto();
+        caseFlagDto3.setFlagCode("PF0015");
+        caseFlagDto3.setCategoryId(2);
+        caseFlagDto3.setCategoryPath("PARTY/REASONABLE ADJUSTMENT");
+        caseFlagDto3.setId(3);
+        caseFlagDto3.setHearingRelevant(true);
+        caseFlagDto3.setRequestReason(false);
+        caseFlagDto3.setValueEn("CHILD OF REASONABLE ADJUSTMENT");
+        caseFlagDto3.setValueCy("");
+        caseFlagDto3.setIsParent(false);
+        caseFlagDto3.setExternallyAvailable(true);
+        caseFlagDto3.setDefaultStatus("Requested");
+        caseFlagDtoList.add(caseFlagDto3);
+        return caseFlagDtoList;
     }
 
     List<CaseFlagDto> getCaseFlagDtoList() {
@@ -128,9 +217,11 @@ class CaseFlagServiceImplTest {
         caseFlagDto1.setValueEn("PARTY");
         caseFlagDto1.setValueCy("");
         caseFlagDto1.setIsParent(true);
+        caseFlagDto1.setFlagCode("CATEGORY");
+        caseFlagDto1.setExternallyAvailable(true);
+        caseFlagDto1.setDefaultStatus("Requested");
 
         var caseFlagDto2 = new CaseFlagDto();
-        caseFlagDto2.setFlagCode("CATEGORY");
         caseFlagDto2.setCategoryId(1);
         caseFlagDto2.setCategoryPath("PARTY");
         caseFlagDto2.setId(2);
@@ -139,6 +230,8 @@ class CaseFlagServiceImplTest {
         caseFlagDto2.setValueEn("REASONABLE ADJUSTMENT");
         caseFlagDto2.setValueCy("");
         caseFlagDto2.setIsParent(true);
+        caseFlagDto2.setExternallyAvailable(true);
+        caseFlagDto2.setDefaultStatus("Active");
 
         var caseFlagDto3 = new CaseFlagDto();
         caseFlagDto3.setFlagCode("FLAG001");
@@ -150,6 +243,8 @@ class CaseFlagServiceImplTest {
         caseFlagDto3.setValueEn("CHILD OF REASONABLE ADJUSTMENT");
         caseFlagDto3.setValueCy("");
         caseFlagDto3.setIsParent(false);
+        caseFlagDto3.setExternallyAvailable(true);
+        caseFlagDto3.setDefaultStatus("Requested");
 
         var caseFlagDto4 = new CaseFlagDto();
         caseFlagDto4.setFlagCode("CATEGORY");
@@ -160,7 +255,9 @@ class CaseFlagServiceImplTest {
         caseFlagDto4.setRequestReason(false);
         caseFlagDto4.setValueEn("CASE");
         caseFlagDto4.setValueCy("");
-        caseFlagDto4.setIsParent(true);
+        caseFlagDto4.setIsParent(false);
+        caseFlagDto4.setExternallyAvailable(true);
+        caseFlagDto4.setDefaultStatus("Active");
 
         var caseFlagDto5 = new CaseFlagDto();
         caseFlagDto5.setFlagCode("CATEGORY");
@@ -172,6 +269,8 @@ class CaseFlagServiceImplTest {
         caseFlagDto5.setValueEn("COMPLEX CASE");
         caseFlagDto5.setValueCy("");
         caseFlagDto5.setIsParent(false);
+        caseFlagDto5.setExternallyAvailable(true);
+        caseFlagDto5.setDefaultStatus("Requested");
 
 
         var caseFlagDtoList = new ArrayList<CaseFlagDto>();
@@ -192,7 +291,7 @@ class CaseFlagServiceImplTest {
         caseFlagDto1.setHearingRelevant(true);
         caseFlagDto1.setRequestReason(false);
         caseFlagDto1.setValueEn("PARTY");
-        caseFlagDto1.setValueCy("");
+        caseFlagDto1.setValueCy("PARTY");
         caseFlagDto1.setIsParent(true);
 
         var caseFlagDto2 = new CaseFlagDto();
@@ -203,8 +302,10 @@ class CaseFlagServiceImplTest {
         caseFlagDto2.setHearingRelevant(true);
         caseFlagDto2.setRequestReason(false);
         caseFlagDto2.setValueEn("Language Interpreter");
-        caseFlagDto2.setValueCy("");
+        caseFlagDto2.setValueCy("Language Interpreter");
         caseFlagDto2.setIsParent(false);
+        caseFlagDto1.setExternallyAvailable(false);
+        caseFlagDto1.setDefaultStatus("Requested");
 
         var caseFlagDtoList = new ArrayList<CaseFlagDto>();
         caseFlagDtoList.add(caseFlagDto1);
@@ -222,20 +323,24 @@ class CaseFlagServiceImplTest {
         caseFlagDto1.setHearingRelevant(true);
         caseFlagDto1.setRequestReason(false);
         caseFlagDto1.setValueEn("PARTY");
-        caseFlagDto1.setValueCy("");
+        caseFlagDto1.setValueCy("PARTY");
         caseFlagDto1.setIsParent(true);
+        caseFlagDto1.setExternallyAvailable(false);
+        caseFlagDto1.setDefaultStatus("Requested");
 
         var caseFlagDto2 = new CaseFlagDto();
         caseFlagDto2.setFlagCode("RA0042");
         caseFlagDto2.setCategoryId(1);
         caseFlagDto2.setCategoryPath("PARTY");
-        caseFlagDto2.setValueCy("");
+        caseFlagDto2.setValueCy("PARTY");
         caseFlagDto2.setId(2);
         caseFlagDto2.setHearingRelevant(true);
         caseFlagDto2.setRequestReason(false);
         caseFlagDto2.setValueEn("Sign Language");
         caseFlagDto2.setValueCy("");
         caseFlagDto2.setIsParent(false);
+        caseFlagDto1.setExternallyAvailable(true);
+        caseFlagDto1.setDefaultStatus("Approved");
 
         var caseFlagDtoList = new ArrayList<CaseFlagDto>();
         caseFlagDtoList.add(caseFlagDto1);
@@ -244,30 +349,75 @@ class CaseFlagServiceImplTest {
         return caseFlagDtoList;
     }
 
-    private List<ListOfValue> getListOfValuesForLanguageInterPreter() {
+    private List<CaseFlagDto> getCaseFlagDtoListWithOther() {
+        var caseFlagDto1 = new CaseFlagDto();
+        caseFlagDto1.setFlagCode("OT0001");
+        caseFlagDto1.setCategoryId(0);
+        caseFlagDto1.setCategoryPath("");
+        caseFlagDto1.setId(1);
+        caseFlagDto1.setHearingRelevant(true);
+        caseFlagDto1.setRequestReason(false);
+        caseFlagDto1.setValueEn("Other");
+        caseFlagDto1.setValueCy("Arall");
+        caseFlagDto1.setIsParent(true);
+        caseFlagDto1.setExternallyAvailable(false);
+        caseFlagDto1.setDefaultStatus("Requested");
+
+        var caseFlagDto2 = new CaseFlagDto();
+        caseFlagDto2.setFlagCode("RA0042");
+        caseFlagDto2.setCategoryId(1);
+        caseFlagDto2.setCategoryPath("PARTY");
+        caseFlagDto2.setValueCy("PARTY");
+        caseFlagDto2.setId(2);
+        caseFlagDto2.setHearingRelevant(true);
+        caseFlagDto2.setRequestReason(false);
+        caseFlagDto2.setValueEn("Sign Language");
+        caseFlagDto2.setValueCy("");
+        caseFlagDto2.setIsParent(false);
+        caseFlagDto1.setExternallyAvailable(true);
+        caseFlagDto1.setDefaultStatus("Approved");
+
+        var caseFlagDtoList = new ArrayList<CaseFlagDto>();
+        caseFlagDtoList.add(caseFlagDto1);
+        caseFlagDtoList.add(caseFlagDto2);
+
+        return caseFlagDtoList;
+    }
+
+
+
+    private List<ListOfValue> getListOfValuesForLanguageInterPreter(boolean isWelshRequired) {
         var list = new ListOfValue();
         list.setId("1");
         list.setKey("EN");
         list.setValue("ENGLISH");
-
+        if (isWelshRequired) {
+            list.setValueCy("CY ENGLISH");
+        } else {
+            list.setValueCy(IGNORE_JSON);
+        }
 
         var listOfValues = new ArrayList<ListOfValue>();
         listOfValues.add(list);
         return listOfValues;
     }
 
-    private List<ListOfValue> getListOfValuesForSignLanguage() {
+    private List<ListOfValue> getListOfValuesForSignLanguage(boolean isWelshRequired) {
         var list = new ListOfValue();
         list.setId("2");
         list.setKey("AF");
         list.setValue("AFRICAN");
-
+        if (isWelshRequired) {
+            list.setValueCy("CY AFRICAN");
+        } else {
+            list.setValueCy(IGNORE_JSON);
+        }
         var listOfValues = new ArrayList<ListOfValue>();
         listOfValues.add(list);
         return listOfValues;
     }
 
-    private void verifyListOfValuesResponse(CaseFlag caseFlag) {
+    private void verifyListOfValuesResponse(CaseFlag caseFlag, boolean isWelshRequired) {
         for (FlagDetail flagDetail : caseFlag.getFlags()
             .get(0).getFlagDetails().get(0).getChildFlags()) {
             if (flagDetail.getListOfValues() != null) {
@@ -275,11 +425,11 @@ class CaseFlagServiceImplTest {
             }
             if (flagDetail.getFlagCode().equals(FLAG_RA0042)) {
                 assertNull(flagDetail.getChildFlags());
-                assertEquals(getListOfValuesForSignLanguage(), flagDetail.getListOfValues());
+                assertEquals(getListOfValuesForSignLanguage(isWelshRequired), flagDetail.getListOfValues());
             }
             if (flagDetail.getFlagCode().equals(FLAG_PF0015)) {
                 assertNull(flagDetail.getChildFlags());
-                assertEquals(getListOfValuesForLanguageInterPreter(), flagDetail.getListOfValues());
+                assertEquals(getListOfValuesForLanguageInterPreter(isWelshRequired), flagDetail.getListOfValues());
             }
         }
         assertEquals(2, caseFlag.getFlags()
