@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentMatchers;
 import org.mockito.InjectMocks;
-import org.mockito.Mock;
+import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.jpa.domain.Specification;
 import uk.gov.hmcts.reform.cdapi.controllers.request.CategoryRequest;
@@ -24,7 +24,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.doReturn;
 import static uk.gov.hmcts.reform.cdapi.helper.CrdTestSupport.buildCategoryRequest;
 
 @ExtendWith(MockitoExtension.class)
@@ -33,7 +33,7 @@ class CrdServiceImplTest {
     @InjectMocks
     CrdServiceImpl crdServiceImpl;
 
-    @Mock
+    @Spy
     ListOfValuesRepository listOfValuesRepository;
 
     @Test
@@ -41,8 +41,9 @@ class CrdServiceImplTest {
         List<ListOfValueDto> listOfValueDtos = new ArrayList<>();
         listOfValueDtos.add(CrdTestSupport.createListOfCategoriesDtoMock("HearingChannel", null,
                                                                          null, null, null));
-        when(listOfValuesRepository.findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any()))
-            .thenReturn(listOfValueDtos);
+        doReturn(listOfValueDtos).when(listOfValuesRepository)
+            .findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any());
+
         CategoryRequest request = buildCategoryRequest("HearingChannel",  null, null,
                                                        null,null, null);
         List<Category> result = crdServiceImpl.retrieveListOfValuesByCategory(request);
@@ -51,6 +52,7 @@ class CrdServiceImplTest {
         Category actualCategory = result.get(0);
         assertEquals(listOfValueDtos.get(0).getCategoryKey().getKey(), actualCategory.getKey());
         assertEquals(listOfValueDtos.get(0).getCategoryKey().getCategoryKey(), actualCategory.getCategoryKey());
+        assertEquals(listOfValueDtos.get(0).getCategoryKey().getServiceId(), actualCategory.getServiceId());
         assertEquals(listOfValueDtos.get(0).getActive(), actualCategory.getActiveFlag());
         assertEquals(listOfValueDtos.get(0).getParentCategory(), actualCategory.getParentCategory());
         assertEquals(listOfValueDtos.get(0).getParentKey(), actualCategory.getParentKey());
@@ -59,8 +61,6 @@ class CrdServiceImplTest {
         assertEquals(listOfValueDtos.get(0).getHintTextCy(), actualCategory.getHintTextCy());
         assertEquals(listOfValueDtos.get(0).getHintTextEn(), actualCategory.getHintTextEn());
         assertEquals(listOfValueDtos.get(0).getLovOrder(), actualCategory.getLovOrder());
-        assertEquals(listOfValueDtos.get(0).getServiceId(), actualCategory.getServiceId());
-        assertEquals(listOfValueDtos.get(0).getActive(), actualCategory.getActiveFlag());
         assertNull(actualCategory.getChildNodes());
     }
 
@@ -72,8 +72,8 @@ class CrdServiceImplTest {
         inactiveCategory.setActive("n");
         listOfValueDtos.add(inactiveCategory);
 
-        when(listOfValuesRepository.findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any()))
-            .thenReturn(listOfValueDtos);
+        doReturn(listOfValueDtos).when(listOfValuesRepository)
+            .findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any());
 
         CategoryRequest request = buildCategoryRequest("HearingChannel", "BBA3", "telephone",
                                            "HearingChannel", "telephone","y");
@@ -83,8 +83,8 @@ class CrdServiceImplTest {
         assertEquals(listOfValueDtos.get(0).getCategoryKey().getKey(), result.get(0).getKey());
         assertEquals(listOfValueDtos.get(0).getCategoryKey().getCategoryKey(), result.get(0).getCategoryKey());
         assertEquals("y", result.get(0).getActiveFlag());
-        assertEquals("y", result.get(1).getActiveFlag());
-        assertEquals("n", result.get(2).getActiveFlag());
+        assertEquals("n", result.get(1).getActiveFlag());
+        assertEquals("y", result.get(2).getActiveFlag());
     }
 
 
@@ -104,8 +104,8 @@ class CrdServiceImplTest {
     @Test
     void retrieveCategoriessByCategoryIdWithChildNodes() {
         List<ListOfValueDto> listOfValueDtos = buildListOfValuesDtos();
-        when(listOfValuesRepository.findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any()))
-            .thenReturn(listOfValueDtos);
+        doReturn(listOfValueDtos).when(listOfValuesRepository)
+            .findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any());
 
         CategoryRequest request = buildCategoryRequest("HearingChannel",  null, null,
                                                        null,null, "y");
@@ -126,8 +126,8 @@ class CrdServiceImplTest {
     @Test
     void retrieveCategoriesByCategoryIdWithNoChildNodes() {
         List<ListOfValueDto> listOfValueDtos = buildListOfValuesDtos();
-        when(listOfValuesRepository.findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any()))
-            .thenReturn(listOfValueDtos);
+        doReturn(listOfValueDtos).when(listOfValuesRepository)
+            .findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any());
 
         CategoryRequest request = buildCategoryRequest("HearingSubChannel",  null, null,
                                            "HearingChannel","telephone", "y");
@@ -149,8 +149,8 @@ class CrdServiceImplTest {
 
     private void verifyWithParams(String isChildRequired, String parentCategory, String parentKey) {
         List<ListOfValueDto> listOfValueDtos = buildListOfValuesDtos();
-        when(listOfValuesRepository.findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any()))
-            .thenReturn(listOfValueDtos);
+        doReturn(listOfValueDtos).when(listOfValuesRepository)
+            .findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any());
 
         CategoryRequest request = buildCategoryRequest("HearingSubChannel", null, null,
                                                        parentCategory, parentKey, isChildRequired);
@@ -167,8 +167,8 @@ class CrdServiceImplTest {
         List<ListOfValueDto> listOfValueDtos = List.of(CrdTestSupport.createListOfCategoriesDtoMock(
             "HearingChannel", "", "", "", ""));
 
-        when(listOfValuesRepository.findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any()))
-            .thenReturn(listOfValueDtos);
+        doReturn(listOfValueDtos).when(listOfValuesRepository)
+            .findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any());
 
         CategoryRequest request = buildCategoryRequest("HearingChannel",  "", "",
                                                        null,"", "");
@@ -185,13 +185,150 @@ class CrdServiceImplTest {
     void shouldThrowNotFoundExceptionWithUnMappedParams() {
         List<ListOfValueDto> listOfValueDtos = new ArrayList<>();
 
-        when(listOfValuesRepository.findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any()))
-            .thenReturn(listOfValueDtos);
+        doReturn(listOfValueDtos).when(listOfValuesRepository)
+            .findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any());
 
         CategoryRequest request = buildCategoryRequest("HearingChannel",  null, null,
                                                        null,null, "n");
         assertThrows(ResourceNotFoundException.class, () -> crdServiceImpl.retrieveListOfValuesByCategory(request));
     }
 
+    @Test
+    void retrieveCategoriesByServiceIdWithNoChildNodes() {
+        List<ListOfValueDto> listOfValueDtos = buildListOfValuesDtos();
+        doReturn(listOfValueDtos).when(listOfValuesRepository)
+            .findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any());
 
+        CategoryRequest request = buildCategoryRequest("HearingChannel",  "BBA3", null,
+                                                       null,null, "n");
+        List<Category> result = crdServiceImpl.retrieveListOfValuesByCategory(request);
+
+        assertNotNull(result);
+        assertEquals(listOfValueDtos.get(0).getCategoryKey().getServiceId(), result.get(0).getServiceId());
+        assertEquals(listOfValueDtos.get(0).getCategoryKey().getCategoryKey(), result.get(0).getCategoryKey());
+        assertEquals(listOfValueDtos.get(0).getActive(), result.get(0).getActiveFlag());
+        assertNull(result.get(0).getChildNodes());
+    }
+
+    @Test
+    void retrieveCategoriesByServiceIdWithChildNodes() {
+        List<ListOfValueDto> listOfValueDtos = buildListOfValuesDtos();
+        doReturn(listOfValueDtos).when(listOfValuesRepository)
+            .findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any());
+
+        CategoryRequest request = buildCategoryRequest("HearingChannel",  "BBA3", null,
+                                                       null,null, "y");
+        List<Category> result = crdServiceImpl.retrieveListOfValuesByCategory(request);
+
+        assertNotNull(result);
+        assertThat(result, hasSize(1));
+        assertEquals(listOfValueDtos.get(0).getCategoryKey().getKey(), result.get(0).getKey());
+        assertEquals(listOfValueDtos.get(0).getCategoryKey().getCategoryKey(), result.get(0).getCategoryKey());
+        assertEquals(listOfValueDtos.get(0).getActive(), result.get(0).getActiveFlag());
+        assertEquals(listOfValueDtos.get(0).getCategoryKey().getServiceId(), result.get(0).getServiceId());
+        assertEquals(listOfValueDtos.get(1).getCategoryKey().getKey(),
+                     result.get(0).getChildNodes().get(0).getKey());
+        assertEquals(listOfValueDtos.get(1).getCategoryKey().getCategoryKey(), result.get(0).getChildNodes().get(0)
+            .getCategoryKey());
+    }
+
+    @NotNull
+    private List<ListOfValueDto> buildListOfValuesDtosWithListOfValuesOrders() {
+        List<ListOfValueDto> listOfValueDtos = new ArrayList<>();
+        listOfValueDtos.add(CrdTestSupport.createListOfCategoriesDtoMock("HearingChannel", "BBA3",
+                                                                         null, null,
+                                                                         "telephone", 1L));
+        listOfValueDtos.add(CrdTestSupport.createListOfCategoriesDtoMock("HearingSubChannel", "BBA3",
+                                                                         "HearingChannel",
+                                                                         "telephone", "telephone", 3L));
+        listOfValueDtos.add(CrdTestSupport.createListOfCategoriesDtoMock("HearingSubChannel", "BBA3",
+                                                                         "HearingChannel",
+                                                                         "telephone", "telephone", 2L));
+        return listOfValueDtos;
+    }
+
+    @Test
+    void retrieveCategoriesByServiceIdWithListOfValuesOrders() {
+        List<ListOfValueDto> listOfValueDtos = buildListOfValuesDtosWithListOfValuesOrders();
+        doReturn(listOfValueDtos).when(listOfValuesRepository)
+            .findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any());
+
+        CategoryRequest request = buildCategoryRequest("HearingChannel",  "BBA3", null,
+                                                       null,null, "n");
+        List<Category> result = crdServiceImpl.retrieveListOfValuesByCategory(request);
+
+        assertNotNull(result);
+        assertEquals(listOfValueDtos.get(0).getCategoryKey().getServiceId(), result.get(0).getServiceId());
+        assertEquals(listOfValueDtos.get(0).getCategoryKey().getCategoryKey(), result.get(0).getCategoryKey());
+        assertEquals(listOfValueDtos.get(0).getActive(), result.get(0).getActiveFlag());
+        assertNull(result.get(0).getChildNodes());
+        assertEquals(listOfValueDtos.get(0).getLovOrder(), result.get(0).getLovOrder());
+        assertEquals(listOfValueDtos.get(2).getLovOrder(), result.get(1).getLovOrder());
+        assertEquals(listOfValueDtos.get(1).getLovOrder(), result.get(2).getLovOrder());
+    }
+
+    @NotNull
+    private List<ListOfValueDto> buildListOfValuesDtosWithDefaultOrdering() {
+        List<ListOfValueDto> listOfValueDtos = new ArrayList<>();
+        final var listOfValueMock1 = CrdTestSupport.createListOfCategoriesDtoMock("HearingChannel", "BBA3",
+                                                                                         null, null,
+                                                                                         "telephone"
+        );
+        listOfValueMock1.setValueEn("third");
+        listOfValueDtos.add(listOfValueMock1);
+
+        final var listOfValueMock2 = CrdTestSupport.createListOfCategoriesDtoMock("HearingSubChannel", "BBA3",
+                                                                                         "HearingChannel",
+                                                                                         "telephone", "telephone"
+        );
+        listOfValueMock2.setValueEn("first");
+        listOfValueDtos.add(listOfValueMock2);
+        final var listOfValueMock3 = CrdTestSupport.createListOfCategoriesDtoMock("HearingSubChannel", "BBA3",
+                                                                                         "HearingChannel",
+                                                                                         "telephone", "telephone"
+        );
+        listOfValueMock3.setValueEn("second");
+        listOfValueDtos.add(listOfValueMock3);
+        return listOfValueDtos;
+    }
+
+    @Test
+    void retrieveCategoriesByServiceIdWithDefaultOrdering() {
+        List<ListOfValueDto> listOfValueDtos = buildListOfValuesDtosWithDefaultOrdering();
+        doReturn(listOfValueDtos).when(listOfValuesRepository)
+            .findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any());
+
+        CategoryRequest request = buildCategoryRequest("HearingChannel",  "BBA3", null,
+                                                       null,null, "n");
+        List<Category> result = crdServiceImpl.retrieveListOfValuesByCategory(request);
+
+        assertNotNull(result);
+        assertEquals(listOfValueDtos.get(1).getCategoryKey().getServiceId(), result.get(1).getServiceId());
+        assertEquals(listOfValueDtos.get(1).getCategoryKey().getCategoryKey(), result.get(1).getCategoryKey());
+        assertEquals(listOfValueDtos.get(1).getActive(), result.get(1).getActiveFlag());
+        assertNull(result.get(1).getChildNodes());
+        assertEquals(listOfValueDtos.get(0).getLovOrder(), result.get(2).getLovOrder());
+        assertEquals(listOfValueDtos.get(1).getLovOrder(), result.get(0).getLovOrder());
+        assertEquals(listOfValueDtos.get(2).getLovOrder(), result.get(1).getLovOrder());
+    }
+
+    @Test
+    void retrieveCategoriesByServiceIdWithDefaultOrderingThrowsNullPointerExceptionForMissingDefault() {
+        List<ListOfValueDto> listOfValueDtos = buildListOfValuesDtosWithDefaultOrdering();
+
+        listOfValueDtos.get(1).setValueEn(null);
+
+        doReturn(listOfValueDtos).when(listOfValuesRepository)
+            .findAll(ArgumentMatchers.<Specification<ListOfValueDto>>any());
+
+        CategoryRequest request = buildCategoryRequest("HearingChannel",  "BBA3", null,
+                                                       null,null, "n");
+
+        final var nullPointerException = assertThrows(NullPointerException.class, () ->
+                                                          crdServiceImpl.retrieveListOfValuesByCategory(request),
+                                                      "Expected NPE to be thrown"
+        );
+
+        assertNotNull(nullPointerException);
+    }
 }
